Add basic test for changing table level settings
The bug here is subtle. The problem is that normalization uses the table
level charset / collation to remove redundant definitions at a column
level.

But if that happens and you then change the table level config, the
normalization on the other schema is different. That subsequently means
it detects a diff when there is none, because the literal column
definition has changed.

The columns look like this then:

```
c: `a` varchar(64) COLLATE latin1_bin
other: `a` varchar(64) CHARACTER SET latin1 COLLATE latin1_bin
```

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>